Basic tracing for not async client
