+++ pom.xml
@@ -34,6 +34,7 @@
         <autorelease>false</autorelease>
 
         <version.mp.config>1.3</version.mp.config>
+        <version.mp.rest.client>1.1</version.mp.rest.client>
         <!-- Versions of API dependencies -->
         <osgi-annotation.version>1.0.0</osgi-annotation.version>
